app/requirements.txt
fastapi
uvicorn
pydantic
redis
pymongo[srv]
requests
munch
colorama
python-dateutil
PyYAML
